requirements.txt
ultralytics
opencv-python
torch
matplotlib
Pillow
